+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/browser/Chromium.kt
@@ -1,10 +1,11 @@
 package com.atlassian.performance.tools.infrastructure.api.browser
 
-import com.atlassian.performance.tools.infrastructure.api.os.Ubuntu
+import com.atlassian.performance.tools.infrastructure.api.browser.chromium.Chromium69
+import com.atlassian.performance.tools.infrastructure.api.browser.chromium.Chromium70
 import com.atlassian.performance.tools.ssh.api.SshConnection
 import java.net.URI
-import java.time.Duration.ofMinutes
 
+@Deprecated(message = "Use version specific Chromium. For example `Chromium69`.")
 class Chromium(private val version: String) : Browser {
 
     companion object {
@@ -20,37 +21,15 @@ class Chromium(private val version: String) : Browser {
 
     init {
         if (!versionToUri.keys.contains(version)) {
-            throw IllegalArgumentException("Chromium version '$version' is not supported. You can use only chromium compatible with chrome driver v2.43." +
-                " Use one of the compatible versions: ${getSupportedVersions().joinToString(", ")}.")
+            throw IllegalArgumentException("Chromium version '$version' is not supported.")
         }
     }
 
     override fun install(ssh: SshConnection) {
-        val ubuntu = Ubuntu()
-        ubuntu.install(
-            ssh,
-            listOf(
-                "unzip",
-                "libx11-xcb1",
-                "libxcomposite1",
-                "libxdamage1",
-                "libxi6",
-                "libxtst6",
-                "libnss3",
-                "libcups2",
-                "libxss1",
-                "libxrandr2",
-                "libasound2",
-                "libpango1.0",
-                "libatk1.0-0",
-                "libatk-bridge2.0",
-                "libgtk-3-0"
-            ),
-            ofMinutes(5)
-        )
-        val uri = versionToUri.get(version)
-        ssh.execute("""wget -q "${uri}" -O chromium.zip""", ofMinutes(1))
-        ssh.execute("unzip chromium.zip")
-        ssh.execute("sudo ln -s `pwd`/chrome-linux/chrome /usr/bin/chrome")
+        when (version) {
+            "69" -> Chromium69().install(ssh)
+            "70" -> Chromium70().install(ssh)
+            else -> throw IllegalArgumentException("Chromium version '$version' is not supported.")
+        }
     }
 }