+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/browser/chromium/Chromium70IT.kt
@@ -1,0 +1,35 @@
+package com.atlassian.performance.tools.infrastructure.api.browser.chromium
+
+import com.atlassian.performance.tools.infrastructure.UbuntuContainer
+import com.atlassian.performance.tools.ssh.api.SshConnection
+import org.hamcrest.Matchers
+import org.junit.Assert
+import org.junit.Test
+
+class Chromium70IT {
+
+    @Test
+    fun shouldInstallBrowser() {
+        UbuntuContainer().run { ssh ->
+            val wasInstalledBefore = isChromiumInstalled(ssh)
+
+            Chromium70().install(ssh)
+
+            val isInstalledAfter = isChromiumInstalled(ssh)
+
+            Assert.assertThat(wasInstalledBefore, Matchers.`is`(false))
+            Assert.assertThat(isInstalledAfter, Matchers.`is`(true))
+        }
+    }
+
+    private fun isChromiumInstalled(ssh: SshConnection): Boolean {
+        val result = ssh
+            .safeExecute("ls -lh /usr/bin/chrome")
+        return result.isSuccessful()
+            .and(
+                result
+                    .output
+                    .contains("/root/chrome-linux/chrome")
+            )
+    }
+}